backend/requirements.txt
# FastAPI framework
fastapi==0.110.0
uvicorn[standard]==0.29.0

# AI model integrations
google-generativeai==0.5.2
openai==1.20.0          # Used for DeepSeek & Grok (OpenAI-compatible APIs)
langchain==0.1.16
langchain-google-genai==1.0.3
langchain-openai==0.1.4
langchain-community==0.0.36

# Database
sqlalchemy==2.0.29
asyncpg==0.29.0          # Async PostgreSQL driver
alembic==1.13.1          # Database migrations
psycopg2-binary==2.9.9   # Sync PostgreSQL driver (for Alembic)

# Redis caching
redis==5.0.3
hiredis==2.3.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Environment & Config
python-dotenv==1.0.1
pydantic==2.6.4
pydantic-settings==2.2.1

# HTTP client
httpx==0.27.0
aiohttp==3.9.4

# Bengali NLP utilities
indic-transliteration==2.3.42

# Rate limiting
slowapi==0.1.9

# Testing
pytest==8.1.1
pytest-asyncio==0.23.6

# Logging & Monitoring
structlog==24.1.0

# Utilities
python-dateutil==2.9.0
